Moving protocol interfaces from common to new protocol package
