+++ orcas/l1only.go
@@ -18,15 +18,16 @@ import (
 	"github.com/netflix/rend/common"
 	"github.com/netflix/rend/handlers"
 	"github.com/netflix/rend/metrics"
+	"github.com/netflix/rend/protocol"
 	"github.com/netflix/rend/timer"
 )
 
 type L1OnlyOrca struct {
 	l1  handlers.Handler
-	res common.Responder
+	res protocol.Responder
 }
 
-func L1Only(l1, l2 handlers.Handler, res common.Responder) Orca {
+func L1Only(l1, l2 handlers.Handler, res protocol.Responder) Orca {
 	return &L1OnlyOrca{
 		l1:  l1,
 		res: res,
